Satisfy §19.10
- Modified `src/simulator/coverage_types.h` because §19.10 declares the covergroup `option` numeric members as `int`, but `CoverOptions` modeled `weight`/`goal`/`at_least`/`auto_bin_max` as `uint32_t`/`double`; switching them to `int32_t` aligns the struct composition with the standard and the five sibling option/type_option structs.
- Modified `src/simulator/coverage.cpp` because the now-signed `auto_bin_max` narrows when brace-initializing the `uint32_t` `CoverPoint::auto_bin_count`, so `AddCoverPoint` needs an explicit `static_cast<uint32_t>` to compile under `-Werror`.
- Modified `test/src/unit/test_simulator_clause_19_10.cpp` because the composition rule requires observing each structure's prescribed member types, so added `static_assert`s locking all six structs' member types and updated the covergroup-option test to signed integer values.

Closes #1559

+++ src/simulator/coverage.cpp
@@ -73,8 +73,14 @@ uint32_t CoverageDB::GroupCount() const {
 }
 
 CoverPoint* CoverageDB::AddCoverPoint(CoverGroup* group, std::string name) {
-  group->coverpoints.push_back(CoverPoint{
-      std::move(name), {}, false, true, 0, 0, group->options.auto_bin_max});
+  group->coverpoints.push_back(
+      CoverPoint{std::move(name),
+                 {},
+                 false,
+                 true,
+                 0,
+                 0,
+                 static_cast<uint32_t>(group->options.auto_bin_max)});
   return &group->coverpoints.back();
 }
